requirements.txt
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
Flask-Login==0.5.0
Flask-WTF==0.15.1
Flask-Migrate==3.1.0
Flask-Mail==0.9.1
Flask-SocketIO==5.1.1
Flask-Babel==2.0.0
Werkzeug==2.0.1
SQLAlchemy==1.4.23
python-dotenv==0.19.0
Pillow==8.3.1
transformers==4.11.3
torch==1.9.0
sentencepiece==0.1.96
protobuf==3.17.3
openai==1.12.0 